fix(price): 현재가 등락률 계산 버그 + 해외 종목 일봉 차트 누락 수정
- StockPriceService의 캐시 미스 폴백(장 마감/주말 등)이 전일종가를
  PreviousCloseCache(항상 "오늘" 기준)로 조회하고 있었는데, 이 폴백은
  latestClose의 거래일이 실제 오늘과 다른 경우에만 실행되므로 previousClose가
  latestClose 자기 자신과 같은 날로 잡혀 등락률이 항상 0%로 계산되는 버그였음 -
  latestClose.getTradeDate() 기준으로 직접 조회하도록 수정
- getChart가 시장 구분 없이 항상 국내 테이블만 조회해 해외 종목 일봉 차트가
  항상 빈 배열로 응답되던 버그 수정 - marketType으로 분기, DailyChartResponse의
  가격 필드를 Long에서 Double로 확대(해외 소수점 가격 지원)

core:test 전체 통과(465개, 실패 0)

+++ backend/core/src/main/java/com/quantlime/price/dto/response/DailyChartResponse.java
@@ -3,12 +3,16 @@ package com.quantlime.price.dto.response;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 
+// open/high/low/close는 Double이다 - 국내(원화)는 정수값이지만 해외(달러)는
+// 소수점 단위(예: $317.31)라 국내/해외 공용 응답 타입을 Long으로 두면 해외
+// 가격의 소수점이 잘린다(OverseasDailyPrice가 별도 엔티티로 분리된 것과
+// 동일한 이유).
 public record DailyChartResponse(
     @JsonFormat(pattern = "yyyy-MM-dd", timezone = "Asia/Seoul") LocalDate tradeDate,
-    Long open,
-    Long high,
-    Long low,
-    Long close,
+    Double open,
+    Double high,
+    Double low,
+    Double close,
     Long volume
 ) {
 }
